requirement.txt
torch>=1.7.0
torchvision>=0.8.1
transformers>=4.5.0
Pillow>=8.1.0
numpy>=1.19.2
matplotlib>=3.4.3
scikit-learn
seaborn
